refactor(mocks): move mocks outside of src folder
