+++ e2e/helpers/itslearning-helper.ts
@@ -1,6 +1,6 @@
 import { Helper } from 'codeceptjs'
 import { Server } from 'http'
-import { ItslearningServer } from '../../src/mocks'
+import { ItslearningServer } from '../../mocks'
 
 class ItslearningHelper extends Helper {
   private itslearningMock: ItslearningServer
