+++ mocks/edusharing/mocked-embed-json/image.ts
