+++ mocks/edusharing/server.ts
@@ -6,9 +6,9 @@ import { imageEmbedJson } from './mocked-embed-json/image'
 import { v4 as uuid_v4 } from 'uuid'
 import * as jose from 'jose'
 import urlJoin from 'url-join'
-import { createAutoFormResponse } from '../../backend/util/create-auto-form-response'
-import { serverLog } from '../../utils/server-log'
-import config from '../../utils/config'
+import { createAutoFormResponse } from '../../src/backend/util/create-auto-form-response'
+import { serverLog } from '../../src/utils/server-log'
+import config from '../../src/utils/config'
 
 export const editorUrl = config.EDITOR_URL
