+++ mocks/itslearning.ts
@@ -2,8 +2,8 @@ import express from 'express'
 import { v4 as uuid_v4 } from 'uuid'
 import * as jose from 'jose'
 import urlJoin from 'url-join'
-import config from '../utils/config'
-import { createAutoFormResponse } from '../backend/util/create-auto-form-response'
+import config from '../src/utils/config'
+import { createAutoFormResponse } from '../src/backend/util/create-auto-form-response'
 
 const itslearningMockDeploymentId = '1'
 const itslearningMockIssuer = 'http://localhost:8101/itslearning'
