+++ mocks/start.ts
@@ -1,11 +1,10 @@
-import { serverLog } from '../utils/server-log'
+import { serverLog } from '../src/utils/server-log'
 import { EdusharingServer } from './edusharing/server'
 import { ItslearningServer } from './itslearning'
 
 const edusharingPort = 8100
-const edusharingServer = new EdusharingServer()
 
-edusharingServer.listen(edusharingPort, () => {
+new EdusharingServer().listen(edusharingPort, () => {
   serverLog('INFO: Mocked version of edusharing is ready.')
   serverLog(
     `Open http://localhost:${edusharingPort}/ to open the Serlo Editor via LTI`
@@ -13,9 +12,8 @@ edusharingServer.listen(edusharingPort, () => {
 })
 
 const itslearningPort = 8101
-const itslearningServer = new ItslearningServer()
 
-itslearningServer.listen(itslearningPort, () => {
+new ItslearningServer().listen(itslearningPort, () => {
   serverLog('INFO: Mocked version of itslearning is ready.')
   serverLog(
     `Open http://localhost:${itslearningPort}/ to open the Serlo Editor via LTI`
